Additional Use requirement
In the sample recipe, it was missing the GridField class itself. Took me awhile to realise I was missing this class myself even with errors that pinged back. Had to match the code I replicated with this to something from the silverstripe omnipay ui module that has a gridfield action provider there.